Move `CommandExt` to its own file
